requirements.txt
annotated-types==0.6.0
appbuilder-sdk==0.7.0
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
future==1.0.0
greenlet==3.0.3
idna==3.7
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.2
orjson==3.10.3
packaging==24.0
pandas==2.2.2
parameterized==0.9.0
parso==0.8.4
platformdirs==4.2.2
prompt-toolkit==3.0.43
proto-plus==1.22.3
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.18.0
pymochow==1.1.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==26.0.3
requests==2.31.0
six==1.16.0
SQLAlchemy==1.4.52
stack-data==0.6.3
tenacity==8.3.0
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
wcwidth==0.2.13
